api: use u32 for price_cents, clamp pagination params
- Narrow price_cents from u64 to u32 through domain, infra, and interface layers
- Clamp page (min 1) and per_page (1..=50) at the interface boundary

+++ api/src/application/product/commands/product_create.rs
@@ -21,7 +21,7 @@ pub struct ProductCreateCommand {
     pub image_url:           Option<String>,
     pub low_stock_threshold: u32,
     pub name:                String,
-    pub price_cents:         u64,
+    pub price_cents:         u32,
     pub stock:               u32,
     pub symbols:             Vec<String>,
     pub unit_of_measure:     String,
